1.1100-B: init-stopped when no app + misc
Refactored the initialization sequence so that if the check for the
'app.psgi' file fails the container will start with the app "stopped"
instead of dying (as if the app had been started with init-stopped).
This allows for the workflow of creating a container first, then
bootstrapping/configuring the app after-the-fact, and is just more
robust in general.

Also added dev feature which will automatically load a dev version of
RapidApp if the repo dir is mounted on /opt/dev/RapidApp. This is not
documented yet (and may never be) because its for internal development
of this module itself and not meant for end-user consumption.

Also now FROM rapi/rapidapp:1.1100-A which adds the 'sqlite' system
package which is needed for most rapidapp.pl bootstrap scenarios