1.2000-D: Add custom ~/.bashrc to auto-load dev RapidApp
Added /root/.bashrc which is loaded when using docker exec and calling
'bash' to get a shell. So far, all this does is check for and source the
RapidApp dev script, if there is a RapidApp repo mounted on
/opt/dev/RapidApp.

Note this is already done in entrypoint.pl for the app. This handles the
exec to bash shell case, too.

--AMEND--

Note: this commit is failing automated build. This is the 3rd attempt
(force push/replace commits). The problem seems to be a Docker Hub bug
which is described here:

https://forums.docker.com/t/automated-docker-build-fails/22831/5

This thread is very recent. One post suggested changing the order of
ADD/COPY commands, which didn't work for me. On this try I moved the
location of the COPY for the new bashrc file to the end (previously I
added it above existing COPY lines). If this doesn't work, this commit
will still be used but I'll have to update with docker push.