1.3302-A: Add (unofficial/undocumented) '/opt/dev/perllib' feature
Just to make developing with possibly several custom/dev perl libraries
more convenient, added a check for the special directory
'/opt/dev/perllib' which when found will recursively find and include
all 'lib/' directories contained within. The idea is that dev modules
can be mounted with the docker -v option within this special path to
have them automatically included. This is no more sophisticated than
setting the 'PERLLIB' env variable when working on a single host/vm. It
is just a way to pass a PERLLIB into the container environment more
easily.

This code was based on the same sort of logic/hook for
'/opt/dev/RapidApp' which has been available for a long time, but only
works for RapidApp. The new '/opt/dev/perllib/' is the same concept but
general purpose.