1.3404: manually installling RapidApp 1.3404
For now we are still based on rapi/rapidapp:1.3400

Also added docker build command reference as comments at the end of the
Dockerfile. Building and publishing of images to Docker Hub has had to
be done manually in this manner ever since automated builds stop being
provided to free accounts. We did apply for their "open source" program
which will continue to be free if accepted. I don't see any reason why
we/RapidApp project wouldn't qualify, but it has taken a while for them
to answer, and frankly the manual process isn't that much of additional
work so at this point I may not even bother to set automated builds back
up, if they ever are free again. For sure they are not worth a paid
subscription